# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SynCloud API",
    "version": "v1",
    "description": "The public SynCloud controller API. The dashboard, synctl and CI all use these endpoints (plan \u00a75.1). Authenticate with the dashboard session cookie, a personal access token (Bearer), or a signed access-key request (SYN1-HMAC-SHA256, plan \u00a77.1)."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "session": []
    },
    {
      "bearerToken": []
    },
    {
      "accessKey": []
    }
  ],
  "paths": {
...
    "/api/v1/projects/{project}/environments/{env}/services": {
      "get": {
        "operationId": "listServices",
        "tags": [
          "services"
        ],
        "summary": "Services in an environment",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "env",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceList"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/middlewares": {
      "get": {
        "operationId": "listAllMiddlewares",
        "tags": [
          "traefik"
        ],
        "summary": "Every middleware preset, and the preset types",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Middleware"
                      }
                    },
                    "types": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/projects/{project}/middlewares": {
      "get": {
        "operationId": "listMiddlewares",
        "tags": [
          "traefik"
        ],
        "summary": "A project's middleware presets",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Middleware"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "operationId": "createMiddleware",
        "tags": [
          "traefik"
        ],
        "summary": "Create a middleware preset and attach it to services",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "config": {
                    "type": "object"
                  },
                  "services": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "name",
                  "type"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Middleware"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/projects/{project}/middlewares/{middleware}": {
      "put": {
        "operationId": "updateMiddleware",
        "tags": [
          "traefik"
        ],
        "summary": "Replace a middleware's settings and attachments (a basic-auth user without a password keeps it)",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "middleware",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "config": {
                    "type": "object"
                  },
                  "services": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "name",
                  "type"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Middleware"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "deleteMiddleware",
        "tags": [
          "traefik"
        ],
        "summary": "Delete a middleware preset",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "middleware",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "syncloud_session"
      },
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Personal access token (syn_pat_\u2026)"
      },
      "accessKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "SYN1-HMAC-SHA256 Credential=<access key id>, Signature=<hex>, with X-Syn-Date and X-Syn-Content-Sha256 headers"
      }
    },
...
    "responses": {
      "Error": {
        "description": "Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "ServiceSpec": {
        "type": "object",
        "required": [
          "image"
        ],
        "properties": {
          "image": {
            "type": "string",
            "example": "nginx:1.27"
          },
          "entrypoint": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "replaces the image ENTRYPOINT"
          },
          "command": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "env": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "sharedEnv": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "readOnly": true,
            "description": "snapshot of the environment's shared variables for this revision (set by the platform; env wins)"
          },
          "ports": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "container": {
                  "type": "integer"
                },
                "protocol": {
                  "type": "string",
                  "enum": [
                    "http",
                    "tcp",
                    "udp"
                  ]
                }
              }
            }
          },
          "resources": {
            "type": "object",
            "properties": {
              "cpu": {
                "type": "number",
                "description": "cores reserved (default 0.1)"
              },
              "memory": {
                "type": "integer",
                "description": "MiB reserved (default 128)"
              },
              "cpuLimit": {
                "type": "number"
              },
              "memoryLimit": {
                "type": "integer",
                "description": "MiB (default 2 \u00d7 memory)"
              }
            }
          },
          "placement": {
            "type": "object",
            "properties": {
              "strategy": {
                "type": "string",
                "enum": [
                  "spread",
                  "binpack"
                ]
              }
            }
          },
          "health": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "http",
                  "tcp",
                  "cmd"
                ]
              },
              "path": {
                "type": "string"
              },
              "port": {
                "type": "string"
              },
              "command": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "interval": {
                "type": "integer"
              },
              "timeout": {
                "type": "integer"
              },
              "retries": {
                "type": "integer"
              },
              "startPeriod": {
                "type": "integer"
              }
            }
          },
          "deployment": {
            "type": "object",
            "properties": {
              "circuitBreaker": {
                "type": "boolean",
                "default": true
              },
              "rollback": {
                "type": "boolean",
                "default": true
              }
            }
          }
        }
      },
...
      "Service": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "project": {
            "type": "string"
          },
          "environment": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "revision": {
            "type": "integer"
          },
          "desiredCount": {
            "type": "integer"
          },
          "running": {
            "type": "integer"
          },
          "pending": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "deleting": {
            "type": "boolean"
          },
          "spec": {
            "$ref": "#/components/schemas/ServiceSpec"
          },
          "endpoints": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "vip": {
            "type": "string",
            "description": "stable virtual IP inside the private network (\u00a78.6)"
          },
          "dnsName": {
            "type": "string",
            "example": "api.production.shop.syncloud.internal"
          },
          "deployment": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Deployment"
              }
            ],
            "nullable": true
          }
        }
      },
      "ServiceList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Service"
            }
          }
        }
      },
...
      "Deployment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "serviceId": {
            "type": "string"
          },
          "fromRevision": {
            "type": "integer"
          },
          "toRevision": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "in_progress",
              "succeeded",
              "failed",
              "rolled_back",
              "superseded",
              "waiting_hook"
            ]
          },
          "failedTasks": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "Middleware": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "project": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "ip-allowlist",
              "rate-limit",
              "basic-auth",
              "redirect-www",
              "cors",
              "security-headers",
              "circuit-breaker",
              "compress",
              "retry"
            ]
          },
          "config": {
            "type": "object",
            "description": "type-specific: rate-limit {average, burst}; basic-auth {users: [{username, password}], realm} (passwords are write-only); ip-allowlist {sourceRange}; security-headers {hsts, hstsSeconds, frameDeny, noSniff, referrerPolicy}; cors {origins, methods, headers, credentials, maxAge}; circuit-breaker {expression}; retry {attempts}; redirect-www and compress {}"
          },
          "services": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "attached services, ENV/NAME: every HTTP route of each uses it"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}